Added improvements to pickup pool and health pickups
Health pickups now only seek the player if needed. That is, if there are other health orbs going towards the player and they will heal the player to full health, then the heal pickup will not seek the player until neccesary. In addition there is now a cap to the pickup pool, which will reuse current pickups if neccessary, prioritizing reusing non seeking player pickups if possible.