Fixed a bug with tier 3 fireball ability in which enemies die before projectile hits them.
Solution: Makes projectile not seeking an enemy if it is currently seeking and the target enemy is null. Happens in update of projectile. Script. In addition tests rigorously for NAN velocity bug found on #6 but could not replicate. Will assume it was fixed/never was an issue since I remember it happening after another bug occured. Will keep an eye out if it ever appears.